Changed some classes of CGS to be struct as they should have been from the beginnning.
- Mode and Point are now structures
- The copy of System.Drawing.Color is not needed anymore the real System.DrawingColor is used instead
- Updated CGS Test Kernel
- Made SVGAII a little more faster (but this not the complete solution)
