--- source/Cosmos.System2/Graphics/KnownColor.cs
@@ -1,206 +1,0 @@
-﻿// Modified from https://github.com/dotnet/corefx/blob/34bd99277390609219bd742ca10b9f246fb81a98/src/System.Drawing.Primitives/src/System/Drawing/KnownColor.cs
-// This enum is internal in .NET Core 2.0, but will be public in .NET Core 2.1
-
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using System.Diagnostics.CodeAnalysis;
-
-namespace System.Drawing
-{
-    [SuppressMessage("Microsoft.Design", "CA1008:EnumsShouldHaveZeroValue")]
-    internal enum KnownColor
-    {
-        // This enum is order dependant!!!
-        //
-        // The value of these known colors are indexes into a color array.
-        // Do not modify this enum without updating KnownColorTable.
-        //
-
-
-        // 0 - reserved for "not a known color"
-
-        // "System" colors
-        ActiveBorder = 1,
-        ActiveCaption,
-        ActiveCaptionText,
-        AppWorkspace,
-        Control,
-        ControlDark,
-        ControlDarkDark,
-        ControlLight,
-        ControlLightLight,
-        ControlText,
-        Desktop,
-        GrayText,
-        Highlight,
-        HighlightText,
-        HotTrack,
-        InactiveBorder,
-        InactiveCaption,
-        InactiveCaptionText,
-        Info,
-        InfoText,
-        Menu,
-        MenuText,
-        ScrollBar,
-        Window,
-        WindowFrame,
-        WindowText,
-
-        // "Web" Colors
-        FirstColor = 0,
-        Transparent = FirstColor,
-        AliceBlue,
-        AntiqueWhite,
-        Aqua,
-        Aquamarine,
-        Azure,
-        Beige,
-        Bisque,
-        Black,
-        BlanchedAlmond,
-        Blue,
-        BlueViolet,
-        Brown,
-        BurlyWood,
-        CadetBlue,
-        Chartreuse,
-        Chocolate,
-        Coral,
-        CornflowerBlue,
-        Cornsilk,
-        Crimson,
-        Cyan,
-        DarkBlue,
-        DarkCyan,
-        DarkGoldenrod,
-        DarkGray,
-        DarkGreen,
-        DarkKhaki,
-        DarkMagenta,
-        DarkOliveGreen,
-        DarkOrange,
-        DarkOrchid,
-        DarkRed,
-        DarkSalmon,
-        DarkSeaGreen,
-        DarkSlateBlue,
-        DarkSlateGray,
-        DarkTurquoise,
-        DarkViolet,
-        DeepPink,
-        DeepSkyBlue,
-        DimGray,
-        DodgerBlue,
-        Firebrick,
-        FloralWhite,
-        ForestGreen,
-        Fuchsia,
-        Gainsboro,
-        GhostWhite,
-        Gold,
-        Goldenrod,
-        Gray,
-        Green,
-        GreenYellow,
-        Honeydew,
-        HotPink,
-        IndianRed,
-        Indigo,
-        Ivory,
-        Khaki,
-        Lavender,
-        LavenderBlush,
-        LawnGreen,
-        LemonChiffon,
-        LightBlue,
-        LightCoral,
-        LightCyan,
-        LightGoldenrodYellow,
-        LightGray,
-        LightGreen,
-        LightPink,
-        LightSalmon,
-        LightSeaGreen,
-        LightSkyBlue,
-        LightSlateGray,
-        LightSteelBlue,
-        LightYellow,
-        Lime,
-        LimeGreen,
-        Linen,
-        Magenta,
-        Maroon,
-        MediumAquamarine,
-        MediumBlue,
-        MediumOrchid,
-        MediumPurple,
-        MediumSeaGreen,
-        MediumSlateBlue,
-        MediumSpringGreen,
-        MediumTurquoise,
-        MediumVioletRed,
-        MidnightBlue,
-        MintCream,
-        MistyRose,
-        Moccasin,
-        NavajoWhite,
-        Navy,
-        OldLace,
-        Olive,
-        OliveDrab,
-        Orange,
-        OrangeRed,
-        Orchid,
-        PaleGoldenrod,
-        PaleGreen,
-        PaleTurquoise,
-        PaleVioletRed,
-        PapayaWhip,
-        PeachPuff,
-        Peru,
-        Pink,
-        Plum,
-        PowderBlue,
-        Purple,
-        Red,
-        RosyBrown,
-        RoyalBlue,
-        SaddleBrown,
-        Salmon,
-        SandyBrown,
-        SeaGreen,
-        SeaShell,
-        Sienna,
-        Silver,
-        SkyBlue,
-        SlateBlue,
-        SlateGray,
-        Snow,
-        SpringGreen,
-        SteelBlue,
-        Tan,
-        Teal,
-        Thistle,
-        Tomato,
-        Turquoise,
-        Violet,
-        Wheat,
-        White,
-        WhiteSmoke,
-        Yellow,
-        YellowGreen,
-        LastColor = YellowGreen,
-
-        // NEW ADDITIONS IN WHIDBEY - DO NOT MOVE THESE UP OR IT WILL BE A BREAKING CHANGE
-        ButtonFace,
-        ButtonHighlight,
-        ButtonShadow,
-        GradientActiveCaption,
-        GradientInactiveCaption,
-        MenuBar,
-        MenuHighlight,
-    }
-}
