+++ source/Cosmos.System2/Graphics/Mode.cs
@@ -8,30 +8,36 @@ namespace Cosmos.System.Graphics
      */
     //public struct Mode : IEquatable<Mode>, IComparable<Mode>
     //[StructLayout(LayoutKind.Explicit, Size = 12)]
-    public class Mode
+    public struct Mode
     {
-        //[FieldOffset(0)]
         int columns;
-        //[FieldOffset(4)]
         int rows;
-        //[FieldOffset(8)]
         ColorDepth color_depth;
+
+        /* Constuctor of our struct */
+        public Mode(int columns, int rows, ColorDepth color_depth)
+        {
+            this.columns = columns;
+            this.rows = rows;
+            this.color_depth = color_depth;
+        }
 
         public bool Equals(Mode other)
         {
-            if ((object)other == null)
-                return false;
-
-            if (this.columns == other.columns && this.rows == other.rows && this.color_depth == other.color_depth)
+            if (columns == other.columns && rows == other.rows && color_depth == other.color_depth)
+            {
                 return true;
+            }
 
             return false;
         }
 
-        public override bool Equals(Object obj)
+        public override bool Equals(object obj)
         {
             if (obj == null)
+            {
                 return false;
+            }
 
             if (!(obj is Mode))
             {
@@ -57,28 +63,23 @@ namespace Cosmos.System.Graphics
         public int CompareTo(Mode other)
         {
             // color_depth has no effect on the orderiring
-            if (this.columns < other.columns && this.rows < other.rows)
+            if (columns < other.columns && rows < other.rows)
+            {
                 return -1;
+            }
 
-            if (this.columns > other.columns && this.rows > other.rows)
+            if (columns > other.columns && rows > other.rows)
+            {
                 return 1;
+            }
 
             // They are effectively Equals
             return 0;
         }
 
-        public static bool operator ==(Mode mode_a, Mode mode_b)
-        {
-            if ((object)mode_a == (object)mode_b)
-                return true;
+        public static bool operator ==(Mode mode_a, Mode mode_b) => mode_a.Equals(mode_b);
 
-            return mode_a.Equals(mode_b);
-        }
-
-        public static bool operator !=(Mode mode_a, Mode mode_b)
-        {
-            return !(mode_a == mode_b);
-        }
+        public static bool operator !=(Mode mode_a, Mode mode_b) => !(mode_a == mode_b);
 
         public static bool operator >(Mode mode_a, Mode mode_b)
         {
@@ -104,9 +105,7 @@ namespace Cosmos.System.Graphics
 
             result = mode_a.CompareTo(mode_b);
 
-            if (result == 0 || result > 0)
-                return true;
-            return false;
+            return (result == 0 || result > 0) ? true : false;
         }
 
         public static bool operator <=(Mode mode_a, Mode mode_b)
@@ -116,17 +115,11 @@ namespace Cosmos.System.Graphics
             result = mode_a.CompareTo(mode_b);
 
             if (result == 0 || result < 0)
+            {
                 return true;
+            }
 
             return false;
-        }
-
-        /* Constuctor of our struct */
-        public Mode(int columns, int rows, ColorDepth color_depth)
-        {
-            this.columns = columns;
-            this.rows = rows;
-            this.color_depth = color_depth;
         }
 
         public int Columns
@@ -153,7 +146,7 @@ namespace Cosmos.System.Graphics
             }
         }
 
-        public override String ToString()
+        public override string ToString()
         {
             return $"{columns}x{rows}@{(int)color_depth}";
         }
