+++ source/Cosmos.System2/Graphics/Point.cs
@@ -1,0 +1,29 @@
+﻿//#define COSMOSDEBUG
+
+namespace Cosmos.System.Graphics
+{
+    public struct Point
+    {
+        public Point(int x, int y)
+        {
+            this.x = x;
+            this.y = y;
+        }
+
+        private int x;
+
+        public int X
+        {
+            get { return x; }
+            set { x = value; }
+        }
+
+        private int y;
+
+        public int Y
+        {
+            get { return y; }
+            set { y = value; }
+        }
+    }
+}
