+++ source/Cosmos.System2/Graphics/SVGAIIScreen.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 using Cosmos.HAL.Drivers.PCI.Video;
+using System.Drawing;
 
 namespace Cosmos.System.Graphics
 {
@@ -43,7 +44,9 @@ namespace Cosmos.System.Graphics
             mSVGAIIDebugger.SendInternal($"Drawing point to x:{x}, y:{y} with {xColor.Name} Color");
             xSVGAIIDriver.SetPixel((uint)x, (uint)y, (uint)xColor.ToArgb());
             mSVGAIIDebugger.SendInternal($"Done drawing point");
-            xSVGAIIDriver.Update(0, 0, (uint)mode.Columns, (uint)mode.Rows);
+            /* No need to refresh all the screen to make the point appear on Screen! */
+            //xSVGAIIDriver.Update((uint)x, (uint)y, (uint)mode.Columns, (uint)mode.Rows);
+            xSVGAIIDriver.Update((uint)x, (uint)y, 1, 1);
         }
 
         public override void DrawArray(Color[] colors, int x, int y, int width, int height)
